requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.2
pillow==10.2.0
sqlparse==0.4.4
